Switch to using System.nanoTime() in main game loop
This change seems to have fixed the frame rate problems present in the
previous commit. I'm guessing that the addition of the CollisionManager
made the frame times go up just enough so that there was aarge
discrepancy between the actual frame times and the millisecond-precision
frame times calculated using System.currentTimeMillis().