Change design of collision-handling system
This commit changes the collision-handling code so that the logic for
handling collisions between each pair of Sprite subclasses is
implemented by a separate class defined for that pair rather than inside
the Sprite subclasses themselves. This change hopefully makes it easier
to write and think about the collision-handling code, since all of the
code for a given interaction goes in one place.