Add Scenes to game engine and game
This commit is a big commit which adds two types of Scenes, changes the game
engine code to use Scenes and allow transitioning between Scenes, and changes
the game code to fit the new Scene architecture.

This commit is a "first draft" of scenes, and there are a few things I want to
work on more:

1) Thinking about how the GameLogic class should relate to the different
GameScene classes. Currently my game's GameLogic class reaches into the current
GameScene to grab the player character's state.

2) Thinking about how Interactables should work when their behavior depends on
the specific GameScene.

3) Performance testing this commit to make sure it didn't hurt the frame rate.